[fix]解决了下注的问题

+++ Server/系统模块/游戏组件/NN/游戏服务器/TableFrameSink.h
@@ -123,7 +123,7 @@ protected:
     BYTE	 m_MaxRatio;
     BYTE	 m_BankerRatio;
     int m_BankerScore;
-    int m_PlayerBankerRatio[NN_GAME_PLAYER];
+    short m_PlayerBankerRatio[NN_GAME_PLAYER];
 
 	NNPlayerBet m_PlayerAllBets[NN_GAME_PLAYER][MAX_STAKE_COUNT];
 	NNPlayerBet m_PlayerBets[NN_GAME_PLAYER];
